distribute

+++ bigcrawler/geospider/control/message_listener.py
@@ -3,10 +3,11 @@ import redis
 
 from geospider.control.message_analyze import Analyze
 from geospider.control.process_controller import ProcessController
+from geospider.utils.mongodb_helper import connect_mongodb, TaskDao
+from geospider.utils.settings_helper import get_attr
 
 
 class MessageListener(object):
-
     channels = []
 
     def __init__(self, host):
@@ -24,29 +25,34 @@ class MessageListener(object):
 
 
 if __name__ == '__main__':
-    listener = MessageListener('127.0.0.1')
-    listener.subscribe('crawler')
-    p = ProcessController()
+    redis_host = get_attr('REDIS_HOST')
+    sub = get_attr('SUBSCRIBE')
+    listener = MessageListener(redis_host)
+    listener.subscribe(sub)
+    db = connect_mongodb()
+    taskdao = TaskDao(db)
+    localhost = get_attr('LOCAL_HOST')
+    p = ProcessController(localhost)
     p.scan()
-    while(True):
+    while (True):
         msg = listener.listen()
         params = Analyze(msg)
         op = params.get('op')
-        print(op)
-        if op=='starttask':
-            taskid = params.get('taskid')
-            status = params.get('status')
-            print(status)
-            if status=='running':
-                p.start(taskid)
-            elif status=='waitting':
-                p.wait(taskid)
-        elif op=='suspendtask':
-            taskid = params.get('taskid')
-            p.suspend(taskid)
-        elif op=='resumetask':
-            taskid = params.get('taskid')
-            p.resume(taskid)
-        elif op=='terminatetask':
-            taskid = params.get('taskid')
-            p.terminate(taskid)
+        taskid = params.get('taskid')
+        task = taskdao.find_by_id(taskid)
+        slave = task['slave']
+        print(slave)
+        if localhost in slave:
+            print(op)
+            if op == 'starttask':
+                status = params.get('status')
+                if status == 'running':
+                    p.start(taskid)
+                elif status == 'waitting':
+                    p.wait(taskid)
+            elif op == 'suspendtask':
+                p.suspend(taskid)
+            elif op == 'resumetask':
+                p.resume(taskid)
+            elif op == 'terminatetask':
+                p.terminate(taskid)
